requirements.txt
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0
pillow==10.2.0
numpy==1.26.3
onnxruntime==1.16.3
albumentations==1.3.1
